Separeted the window opening dealing code in another .js
